added animal_location entity and refactoring all

+++ src/main/java/pl/coderslab/Animal/AnimalController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import pl.coderslab.User.User;
 import pl.coderslab.User.UserDao;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Slf4j
@@ -23,19 +24,29 @@ public class AnimalController {
     }
 
     @GetMapping(value = "/animal/add/form")
-    public String showAddForm(Model model) {
-        List<User> users = userDao.findAllUsers();
-        model.addAttribute("users", users);
+    public String showAddForm(Model model, HttpSession session) {
         model.addAttribute("animal", new Animal());
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            model.addAttribute("loggedInUser", loggedInUser);
+        } else {
+            return "redirect:/login";
+        }
 //        List<String> classis = List.of("FISH", "AMPHIBIAN", "REPTILE", "BIRD", "MAMMAL");
 //        model.addAttribute("classis", classis);
         return "addAnimal";
     }
 
     @PostMapping(value = "/animal/add/form")
-    public String processAddAnimal(@ModelAttribute Animal animal) {
-        animalDao.saveAnimal(animal);
-        return "redirect:/animal/list";
+    public String processAddAnimal(@ModelAttribute Animal animal, HttpSession session) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            animal.setUserId(loggedInUser.getId());
+            animalDao.saveAnimal(animal);
+            return "redirect:/animal/list";
+        } else {
+            return "redirect:/login";
+        }
     }
 
     @ModelAttribute("classis")
@@ -44,30 +55,53 @@ public class AnimalController {
     }
 
     @GetMapping(value = "/animal/edit/form/{id}")
-    public String editAnimalForm(@PathVariable Long id, Model model) {
+    public String editAnimalForm(@PathVariable Long id, Model model, HttpSession session) {
         Animal animal = animalDao.findAnimalById(id);
         model.addAttribute("animal", animal);
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            model.addAttribute("loggedInUser", loggedInUser);
+        } else {
+            return "redirect:/login";
+        }
         return "editAnimal";
     }
 
-
     @PostMapping(value = "/animal/edit")
-    public String processEditAnimal(@ModelAttribute Animal animal) {
-        animalDao.updateAnimal(animal);
-        return "redirect:/animal/list";
+    public String processEditAnimal(@ModelAttribute Animal animal, HttpSession session) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            animal.setUserId(loggedInUser.getId());
+            animalDao.updateAnimal(animal);
+            return "redirect:/animal/list";
+        } else {
+            return "redirect:/login";
+        }
     }
 
     @GetMapping("/animal/delete/form/{id}")
-    public String deleteAnimalForm(@PathVariable Long id, Model model) {
+    public String deleteAnimalForm(@PathVariable Long id, Model model, HttpSession session) {
         Animal animal = animalDao.findAnimalById(id);
         model.addAttribute("animal", animal);
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            model.addAttribute("loggedInUser", loggedInUser);
+        } else {
+            return "redirect:/login";
+        }
         return "deleteAnimal";
     }
 
+    //@Transactional
     @PostMapping(value = "/animal/delete")
-    public String processDeleteAnimal(@RequestParam Long id) {
-        animalDao.deleteAnimalById(id);
-        return "redirect:/animal/list";
+    public String processDeleteAnimal(@RequestParam Long id, HttpSession session) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            animalDao.deleteAnimalById(id);
+            return "redirect:/animal/list";
+        } else {
+            return "redirect:/login";
+        }
     }
 
     @GetMapping("/animal/list")
