+++ src/main/java/pl/coderslab/Location/LocationController.java
@@ -1,13 +1,13 @@
 package pl.coderslab.Location;
 
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.Hibernate;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import pl.coderslab.User.User;
 import pl.coderslab.User.UserDao;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Slf4j
@@ -26,83 +26,78 @@ public class LocationController {
         this.userDao = userDao;
     }
 
-//    @RequestMapping("/location/add")
-//    @ResponseBody
-//    public String addLocation(@RequestParam Long id, @RequestParam String name) {
-//        Location location = new Location();
-//        location.setName("locationName");
-//        location.setBiome("biome");
-//        locationDao.saveLocation(location);
-//        return "Dodano lokalizację " + location.getName();
-//    }
-
-//    @RequestMapping("/location/get")
-//    @ResponseBody
-//    public String getLocation(@RequestParam Long id) {
-//        Location location = locationDao.findLocationById(id);
-//        return location.toString();
-//    }
-//
-//    @RequestMapping("/location/update/{id}")
-//    @ResponseBody
-//    public String updateLocation(@PathVariable Long id) {
-//        Location location = locationDao.findLocationById(id);
-//        locationDao.updateLocation(location);
-//        return location.toString();
-//    }
-
-//    @RequestMapping("location/delete/{id}")
-//    @ResponseBody
-//    public String deleteLocationById(@PathVariable Long id) {
-//        Location location = locationDao.findLocationById(id);
-//        locationDao.deleteLocationById(location.getId());
-//        return "Deleted" + location;
-//    }
-
-
-
-
     @GetMapping(value = "/location/add/form")
-    public String showAddLocationForm(Model model) {
-        List<User> users = userDao.findAllUsers();
-        model.addAttribute("users", users);
+    public String showAddLocationForm(Model model, HttpSession session) {
         model.addAttribute("location", new Location());
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            model.addAttribute("loggedInUser", loggedInUser);
+        } else {
+            return "redirect:/login";
+        }
         return "addLocation";
     }
 
     @PostMapping(value = "/location/add/form")
-    public String processAddLocation(@ModelAttribute Location location) {
+    public String processAddLocation(@ModelAttribute Location location, HttpSession session) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            location.setUserId(loggedInUser.getId());
         locationDao.saveLocation(location);
         return "redirect:/location/list";
+        } else {
+            return "redirect:/login";
+        }
     }
 
     @GetMapping(value = "/location/edit/form/{id}")
-    public String editLocationForm(@PathVariable Long id, Model model) {
+    public String editLocationForm(@PathVariable Long id, Model model, HttpSession session) {
         Location location = locationDao.findLocationById(id);
         model.addAttribute("location", location);
-        Hibernate.initialize(location.getObservations());
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            model.addAttribute("loggedInUser", loggedInUser);
+        } else {
+            return "redirect:/login";
+        }
         return "editLocation";
     }
 
     @PostMapping(value = "/location/edit")
-    public String processEditLocation(@ModelAttribute Location location) {
+    public String processEditLocation(@ModelAttribute Location location, HttpSession session) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            location.setUserId(loggedInUser.getId());
         locationDao.updateLocation(location);
         return "redirect:/location/list";
+        } else {
+            return "redirect:/login";
+        }
     }
 
     @GetMapping("/location/delete/form/{id}")
-    public String deleteLocationForm(@PathVariable Long id, Model model) {
+    public String deleteLocationForm(@PathVariable Long id, Model model, HttpSession session) {
         Location location = locationDao.findLocationById(id);
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
+            model.addAttribute("loggedInUser", loggedInUser);
+        } else {
+            return "redirect:/login";
+        }
         model.addAttribute("location", location);
         return "deleteLocation";
     }
 
     @PostMapping(value = "/location/delete")
-    public String processDeleteLocation(@RequestParam Long id) {
+    public String processDeleteLocation(@RequestParam Long id, HttpSession session) {
+        User loggedInUser = (User) session.getAttribute("loggedInUser");
+        if (loggedInUser != null) {
         locationDao.deleteLocationById(id);
         return "redirect:/location/list";
+        } else {
+            return "redirect:/login";
+        }
     }
-
 
     @GetMapping("/location/list")
     public String showLocationsList(Model model) {
